requirements.txt
python=3.8 
pytorch==1.8.0
torchvision==0.9.0 
torchaudio==0.8.0
timm==0.3.2